feat(shared,backend): Fix telemetry notice emission
