+++ packages/shared/package.json
@@ -76,7 +76,6 @@
     "clean": "rimraf ./dist",
     "dev": "tsup --watch",
     "dev:publish": "npm run dev -- --env.publish",
-    "postinstall": "node ./scripts/postinstall.mjs",
     "lint": "eslint src/",
     "lint:attw": "attw --pack .",
     "lint:publint": "publint",
